Only embed GL/GLES shaders when we need them
