+++ internal/painter/gl/shaders.go
@@ -1,3 +1,5 @@
+//go:build (!gles && !arm && !arm64 && !android && !ios && !mobile && !test_web_driver && !wasm) || (darwin && !mobile && !ios && !wasm && !test_web_driver)
+
 package gl
 
 import _ "embed"
@@ -9,39 +11,32 @@ var (
 	//go:embed shaders/line.vert
 	shaderLineVert []byte
 
-	//go:embed shaders/line_es.frag
-	shaderLineesFrag []byte
-
-	//go:embed shaders/line_es.vert
-	shaderLineesVert []byte
-
 	//go:embed shaders/rectangle.frag
 	shaderRectangleFrag []byte
 
 	//go:embed shaders/rectangle.vert
 	shaderRectangleVert []byte
 
-	//go:embed shaders/rectangle_es.frag
-	shaderRectangleesFrag []byte
-
-	//go:embed shaders/rectangle_es.vert
-	shaderRectangleesVert []byte
-
 	//go:embed shaders/round_rectangle.frag
 	shaderRoundrectangleFrag []byte
-
-	//go:embed shaders/round_rectangle_es.frag
-	shaderRoundrectangleesFrag []byte
 
 	//go:embed shaders/simple.frag
 	shaderSimpleFrag []byte
 
 	//go:embed shaders/simple.vert
 	shaderSimpleVert []byte
-
-	//go:embed shaders/simple_es.frag
-	shaderSimpleesFrag []byte
-
-	//go:embed shaders/simple_es.vert
-	shaderSimpleesVert []byte
 )
+
+func shaderSourceNamed(name string) ([]byte, []byte) {
+	switch name {
+	case "line":
+		return shaderLineVert, shaderLineFrag
+	case "simple":
+		return shaderSimpleVert, shaderSimpleFrag
+	case "rectangle":
+		return shaderRectangleVert, shaderRectangleFrag
+	case "round_rectangle":
+		return shaderRectangleVert, shaderRoundrectangleFrag
+	}
+	return nil, nil
+}
